+++ src/main/java/com/test/automation/sdk/utility/PropertiesReader.java
@@ -1,6 +1,6 @@
 package com.test.automation.sdk.utility;
 
-import com.test.automation.sdk.testbase.SdkConfig;
+import com.test.automation.sdk.config.SdkConfig;
 
 import java.io.File;
 import java.io.FileInputStream;